src/third_party/blink/web_tests/fast/forms/file/file-input-click.html

<!DOCTYPE html>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/choosers/file_chooser.mojom.js"></script>
<script src="../resources/mock-file-chooser.js"></script>
<title>INPUT TYPE=FILE and click events</title>
<p>
  Ensure clicking on an INPUT element with TYPE=FILE launches a file
  chooser. Automated test passes if 'FileChooser: opened' was
  logged.
</p>
<input type="file">
<script>
if (testRunner && eventSender) {
  testRunner.dumpAsText();
  const file = document.querySelector('input');
  const x = file.offsetLeft + file.offsetWidth / 2;
  const y = file.offsetTop + file.offsetHeight / 2;

  testRunner.waitUntilDone();
  mockFileChooserFactory.addEventListener('open', e => testRunner.notifyDone());

  eventSender.mouseMoveTo(x, y);
  eventSender.mouseDown();
  eventSender.mouseMoveTo(x, y);
  eventSender.mouseUp();
}
</script>
