src/third_party/blink/web_tests/fast/forms/file/file-input-click-expected.txt

CONSOLE MESSAGE: line 5: FileChooser: opened; mode=kOpen
Ensure clicking on an INPUT element with TYPE=FILE launches a file chooser. Automated test passes if 'FileChooser: opened' was logged.
